Extract functionality to manually book a bed
These test will be running against the dev env DB which is cleared
on every deploy of the API.
We can’t guarantee that the DB will have bookings against beds so before
cancelling, marking an arrival or non-arrival we need to ensure there
is a booking. Because we will be repeating this functionality it’s
best to extract it to seperate functions.