Update Apply/Assess tests
This fetches the ID from the Application on submit, so we can follow
through the application, and use the data attributes to ensure we have
the right application.

I’ve also added a conditional to allow us to create an application
without a date, which we’ll use in the next commit.