Add tests for user management feature
I have added one test here to test all the functionality in a light-touch manner.
It is slightly risky to delete a user here but I have opted to add them first
so that if the test fails at any point then when it is run again it should work.
There is no issue with adding a user that already exists - the functionality works the same.
I have opted to use JOSEPHHOLLINSHEAD just in case the user is deleted and we are unable to recover
the account for some unknown reason.
For the search functionality we assert a user that appears later in the list (so they
are not visible in the first page of results) isnt visible. We then search for the user and ensure
they are visible.