Use stored user name for `setRoles`
Previously this step used the test user name to edit the user roles before performing a task - however, this user does not exist in the local environment. Replacing the string value here with the stored `user`, which is used in different steps.

This will make it easier to run the same tests against a local instance of the app.

Add user object structure to dev setup, as we are not currently storing the display name in environment variables